# llama.swiftui : add bench functionality