--- a/examples/llama.swiftui/.gitignore
+++ b/examples/llama.swiftui/.gitignore
@@ -1,1 +1,2 @@
 xcuserdata
+xcshareddata