ref #4358 

Some initial runs on iPhone 13 Mini using TinyLlama 1.1B

https://huggingface.co/TinyLlama/TinyLlama-1.1B-intermediate-step-1195k-token-2.5T

```bash
# clone model
git clone https://huggingface.co/TinyLlama/TinyLlama-1.1B-intermediate-step-1195k-token-2.5T

# convert to GGUF F16
python3 convert.py ./TinyLlama-1.1B-intermediate-step-1195k-token-2.5T/ --outfile models/tinyllama-1b/ggml-model-f16.gguf --outtype f16

# quantize to Q8_0 and Q4_0
./quantize models/tinyllama-1b/ggml-model-f16.gguf models/tinyllama-1b/ggml-model-q8_0.gguf q8_0
./quantize models/tinyllama-1b/ggml-model-f16.gguf models/tinyllama-1b/ggml-model-q4_0.gguf q4_0

# bench
make clean && make -j llama-bench && ./llama-bench \
  -m ./models/tinyllama-1b/ggml-model-f16.gguf  \
  -m ./models/tinyllama-1b/ggml-model-q8_0.gguf \
  -m ./models/tinyllama-1b/ggml-model-q4_0.gguf \
  -p 512 -n 128 -ngl 99 2> /dev/null
```

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 1B mostly Q8_0           |  1.09 GiB |     1.10 B | Metal      |  99 | pp 512     |    397.35 |
| llama 1B mostly Q8_0           |  1.09 GiB |     1.10 B | Metal      |  99 | tg 128     |     23.86 |
| llama 1B mostly Q4_0           |  606.53 MiB |    1.10 B | Metal      |  99 | pp 512     |    408.00 |
| llama 1B mostly Q4_0           |  606.53 MiB |    1.10 B | Metal      |  99 | tg 128     |     39.70 |

TODOs:

- [x] Compute result uncertainty in a similar way as `llama-bench`
- [x] Output markdown results that can be copy-pasted
- [x] Download model in-app? not sure how to do that?
- [x] Text generation should terminate after 64 tokens
- [x] Option to stop benchmark
      Sometimes when the model is too big, the phone can get completely bricked while doing the benchmark. Should add "Stop" button + timing to stop benchmarking after more than 1 minute has passed

<img src="https://github.com/ggerganov/llama.cpp/assets/1991296/7736fed7-b650-4a0a-8bfb-df81704e03d8" width="200"></img>