@jhen0409 2023-12-16
> [ ] Download model in-app? not sure how to do that?

I can help to do this later if not done yet. Would you expected to download multiple models? It will be help with benchmarking in one run.