@bachittle 2023-12-16
I think using this would allow functionality to download from a huggingface URL directly into the app: https://developer.apple.com/documentation/foundation/url_loading_system/downloading_files_from_websites