@ggerganov 2023-12-16
> Would you expected to download multiple models?

I imagined the most trivial functionality would be to always have a single model stored on the device and have a list of buttons:

- TinyLlama Q4_0 (600 MiB)
- TinyLlama Q8_0 (1.09 GiB)
- Phi-2 Q4_0 (.. GiB)
- etc.

When you click on a button, it erases the currently stored model and downloads the selected model from HF in place of it.

The reason to do it like this is that it can become complicated with managing the storage, free space, what is downloaded and what is not downloaded, etc. But if you think that it is manageable and the code would not become too cumbersome, then we can download and store multiple models. I just don't have enough Swift experience to evaluate how difficult this would be to implement and I'm mostly hacking on this example with the help of Copilot.