@jhen0409 2023-12-17
I created a simple download button to do that, it can show progress / abort download and init a llama context. Currently the button does not delete the existing one when performing the download, I added an another clear button.

Currently the models only added [TheBloke/TinyLlama-1.1B-1T-OpenOrca-GGUF](https://huggingface.co/TheBloke/TinyLlama-1.1B-1T-OpenOrca-GGUF) q4_0 and q8_0, we can easily edit it