@ggerganov 2023-12-17
Is there a way to write the downloaded model file in chunks so that we reduce memory usage?

Currently, on my device when I download the Q8_0 model, it crashes during writing the file because I think it runs out of memory. The reason seems to be that in order to write the data, we make a copy (I think)

If I add the model manually (i.e. without going though download) it works