@ggerganov 2023-12-17
> Is there a way to write the downloaded model file in chunks so that we reduce memory usage?

Well, thank you local Mixtral 8x7B 😄 

```bash
./main -m models/mixtral-instruct-8x7b-fast/ggml-model-q8_0.gguf -e -f task.txt --repeat_penalty 1.0 --color --temp 0.0 -c 4096 -n -1 --no-penalize-nl --verbose-prompt
```

<details>
<summary>Answer via `llama.cpp main`</summary>

sampling: 
	repeat_last_n = 64, repeat_penalty = 1.000, frequency_penalty = 0.000, presence_penalty = 0.000
	top_k = 40, tfs_z = 1.000, top_p = 0.950, min_p = 0.050, typical_p = 1.000, temp = 0.000
	mirostat = 0, mirostat_lr = 0.100, mirostat_ent = 5.000
sampling order: 
CFG -> Penalties -> top_k -> tfs_z -> typical_p -> top_p -> min_p -> temp 
generate: n_ctx = 4096, n_batch = 512, n_predict = -1, n_keep = 0


 [INST] How to improve this SwiftUI function to use less memory?

```swift
    private func download() {
        status = "downloading"
        print("Downloading model \(modelName) from \(modelUrl)")
        downloadTask = URLSession.shared.dataTask(with: URL(string: modelUrl)!) { data, response, error in
            if let error = error {
                print("Error: \(error.localizedDescription)")
                return
            }

            guard let response = response as? HTTPURLResponse, (200...299).contains(response.statusCode) else {
                print("Server error!")
                return
            }

            if let data = data {
                do {
                    print("Writing to \(filename)")
                    let fileURL = DownloadButton.getFileURL(filename: filename)
                    try data.write(to: fileURL)

                    llamaState.cacheCleared = false

                    status = "downloaded"
                    try llamaState.loadModel(modelUrl: fileURL)
                } catch let err {
                    print("Error: \(err.localizedDescription)")
                }
            }
        }
        observation = downloadTask?.progress.observe(\.fractionCompleted) { progress, _ in
            self.progress = progress.fractionCompleted
        }
        downloadTask?.resume()
    }
```

Currently, when it downloads a 1GB file, it uses 2GB of memory which is very wasteful. I would like to avoid the copy of the data. [/INST] To avoid copying the data and using less memory while downloading large files in SwiftUI, you can write the data directly to the file as it is being downloaded, instead of loading the entire data into memory before writing it to the file. Here's an updated version of your function that does this:

```swift
private func download() {
    status = "downloading"
    print("Downloading model \(modelName) from \(modelUrl)")
    guard let url = URL(string: modelUrl) else { return }
    let fileURL = DownloadButton.getFileURL(filename: filename)

    downloadTask = URLSession.shared.downloadTask(with: url) { temporaryURL, response, error in
        if let error = error {
            print("Error: \(error.localizedDescription)")
            return
        }

        guard let response = response as? HTTPURLResponse, (200...299).contains(response.statusCode) else {
            print("Server error!")
            return
        }

        do {
            if let temporaryURL = temporaryURL {
                try FileManager.default.copyItem(at: temporaryURL, to: fileURL)
                print("Writing to \(filename) completed")

                llamaState.cacheCleared = false

                status = "downloaded"
                try llamaState.loadModel(modelUrl: fileURL)
            }
        } catch let err {
            print("Error: \(err.localizedDescription)")
        }
    }

    observation = downloadTask?.progress.observe(\.fractionCompleted) { progress, _ in
        self.progress = progress?.fractionCompleted
    }

    downloadTask?.resume()
}
```

This version uses the `downloadTask(with:)` method from `URLSession`, which downloads the file directly to the specified `fileURL`. This way, you avoid loading the entire file into memory, and the memory usage should be much lower. [end of text]

llama_print_timings:        load time =    1615.12 ms
llama_print_timings:      sample time =      13.26 ms /   461 runs   (    0.03 ms per token, 34768.84 tokens per second)
llama_print_timings: prompt eval time =    3456.45 ms /   384 tokens (    9.00 ms per token,   111.10 tokens per second)
llama_print_timings:        eval time =   12839.38 ms /   460 runs   (   27.91 ms per token,    35.83 tokens per second)
llama_print_timings:       total time =   16372.73 ms
</details>