@ggerganov 2023-12-17
iPhone 13 Mini

Loading model...
Could not locate model
Loading model...
Loaded model tinyllama-1.1b-1t-openorca.Q4_0.gguf

Running benchmark...
Model info: llama 1B Q4_0
Heat up time: 0.630616959 seconds, please wait...
| model | size | params | backend | test | t/s |
| --- | --- | --- | --- | --- | --- |
| llama 1B Q4_0 | 0.59 GiB | 1.10 B | Metal | pp 512 | 405.30 ± 7.26 |
| llama 1B Q4_0 | 0.59 GiB | 1.10 B | Metal | tg 128 | 39.03 ± 0.08 |
Loading model...
Loaded model tinyllama-1.1b-1t-openorca.Q8_0.gguf

Running benchmark...
Model info: llama 1B Q8_0
Heat up time: 1.659307333 seconds, please wait...
| model | size | params | backend | test | t/s |
| --- | --- | --- | --- | --- | --- |
| llama 1B Q8_0 | 1.09 GiB | 1.10 B | Metal | pp 512 | 411.16 ± 6.22 |
| llama 1B Q8_0 | 1.09 GiB | 1.10 B | Metal | tg 128 | 24.12 ± 0.04 |

![image](https://github.com/ggerganov/llama.cpp/assets/1991296/1311ed73-166a-454b-95e8-b1beea77aa29)